Metrics code style cleanup
